+++ src/components/Modal/Modal.tsx
@@ -1,28 +1,13 @@
-import React, { useRef, useEffect } from 'react';
+import React from 'react';
 import './Modal.css';
+import useModal from './useModal';
 
 export interface ModalProps {
   children?: React.ReactNode;
 }
 
 function Modal(props: ModalProps) {
-  const modalRef = useRef<HTMLDialogElement>(null);
-
-  useEffect(() => {
-    if (modalRef.current) modalRef.current.showModal();
-  }, [props]);
-
-  const closeModal = () => {
-    if (modalRef.current) {
-      modalRef.current.close();
-    }
-  };
-
-  const onBackdropClick = (event: React.MouseEvent<HTMLDialogElement>) => {
-    if (event.target === modalRef.current) {
-      closeModal();
-    }
-  };
+  const { modalRef, closeModal, onBackdropClick } = useModal(props);
 
   return (
     <dialog ref={modalRef} onClick={onBackdropClick}>
@@ -31,7 +16,7 @@ function Modal(props: ModalProps) {
         <button
           className="button button--primary text-caption"
           id="restaurant-detail-modal-close-button"
-          onClick={() => closeModal()}
+          onClick={closeModal}
         >
           닫기
         </button>
